New functionality and parameterized functions
Added beautifulsoup functionality. Added ability to find and list html comments and links during in-line processing. Added ability to select whether or not to write the files and added tmp folder cleanup functionality after zipping the file. Added parameter to set max files to retrieve for testing or to limit total retrievals.